+++ examples/plugin-react/src/index.ts
@@ -1,3 +1,4 @@
+import path from 'path';
 import { IContext, PluginOptions } from '@alicloud/console-toolkit-core';
 
 import getProdConfig from './config/prod.js';
@@ -6,6 +7,19 @@ import build from './build.js';
 import start from './server.js';
 
 export default function (context: IContext, options: PluginOptions) {
+  context.registerCommand('init <projectName>', {
+    description: 'init',
+    usage: '',
+  }, async (_options, args) => {
+    const [projectName] = args || [];
+
+    context.dispatchAPI('generate', {
+      name: '@ali/xconsole-scaffold',
+      dir: path.resolve(context.cwd, projectName),
+      registry: 'http://registry.npm.alibaba-inc.com',
+    });
+  });
+
   context.registerCommand('build', {
     description: 'build',
     usage: '',